ssh-key: verify existance before use
Instead of assuming that everyone has an ssh key generated and it is
always stored in '~/.ssh/id_rsa.pub', instead ask the user which
key they would like to use. Then verify this key exists before going any
further in the playbook.

This is much better than getting about one quarter the way through the
playbook only to receive a lookup exception with no context that their
ssh key might be the problem.

Change-Id: Ia77287762034889bd2318ba6d2c414eae0a55420
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>