shut down after installation, so that virt-install knows when it is finished
Currently (at least on my F25 system) the virt-install command gets
stuck forever. It's waiting for virt-viewer to terminate, as an
indication that the installation is finished, but after Anaconda reboots
the machine virt-viewer simply keeps running while the machine boots
back up.

Instead of rebooting after the installation, we want the machine to
power off so that virt-viewer exits so that virt-install knows the
installation is finished and the command will exit. Then we can start
the machine back up and let the playbook continue.

Change-Id: Ibc252c638e2010d64a976da8a74b52679587533b