Use beaker in a box for a development sandbox
This patch introduces a new playbook which can be used to deploy a basic
development sandbox. It re-uses roles from beaker-in-a-box to achieve
semi-consistency with the Beaker in a box environment.

This patch makes adjustments on how the two playbooks read in variables. Due to
the fact that we run a first play against localhost, group vars are really not
the ideal place to share variables among the hosts here. The localhost can not
really be grouped and shared between sandbox and beaker-in-a-box. If I group it
with the sandbox, the variables are available for the localhost play and the
sandbox play. But running beaker-in-a-box fails, because it accesses the
variables for localhost for the sandbox (same group!).

Instead of grouping them with hosts, I've moved the variable definitions into
their own variable files included in the playbooks to avoid above scenario.

Bug: BKR-3612
Change-Id: I2f4a0edb1e14c2afd6a3dece05863a3ccb713cee