Verify variables before running playbook
Instead of blindly assuming that all variables point to correct values, add a
pre-flightcheck to the playbook and bail out if not all needed variables are
correctly setup. In case of failure, point to our developer documentation.

This patch includes a bit of restructuring of variables as well. I've changed
how we determine the user who's running the playbook, since the ansible_user
variables typically point to 'root' since the tasks are executed as root on the
remote system.

Change-Id: I685a41829d73e2a7020ce8713b2d92d807f2e139