Perform the inventory scan as 'root'
The step in our test playbook to verify that the Beaker-in-a-box is correctly
setup, performs a system scan monitored by the Beaker client. The
beaker_server_lc.beaker host has been provisioned without an additional user and
the user who uses the host will log in as root.

The test playbook which issues the bkr system scan has always been running as
root. However this is not really reflecting reality, since we discourage running
these playbooks as the root user, but rather with a normal unprivileged user
which gains root privileges through sudo.

In order to reflect the normal use case, we'll be changing the test which runs
the playbooks to run as a normal user. We need to reflect in our test playbook
however, that the playbook needs to login as root, since that's the only user
which exists on the newly provisioned beaker_server_lc.beaker.

Change-Id: I6c734f5acb79a3d681eaa8968b3657e012cff00e