Check for a user_key instead of user_key_path
The task failed if a user accepts the default user_key_path variable which
defaults to "~/.ssh/id_rsa.pub". The problem is that ansible will not expand the
variable and the task fails. Instead, the playbook sets the user_key, which is
the evaluated user_key_path variable. We should check if this is not empty
instead of the not expanding user_key_path string.

Change-Id: I8be945c922c5b4d9a4b674d1c96350c3cebe3d04